rename: file to main_ocr_test_screen
